+++ spec/auction_fun_core_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe AuctionFunCore do
   it "has a version number" do
-    expect(AuctionFunCore::VERSION).to eq("0.8.7")
+    expect(AuctionFunCore::VERSION).to eq("0.8.8")
   end
 end